Update OpenSuggestionCommand to use BlockTreeItem
