+++ src/main/java/com/notably/logic/suggestion/commands/OpenSuggestionCommand.java
@@ -8,8 +8,9 @@ import java.util.stream.Collectors;
 
 import com.notably.commons.core.path.AbsolutePath;
 import com.notably.logic.suggestion.SuggestionCommand;
-import com.notably.logic.suggestion.commands.treetest.Node;
 import com.notably.model.Model;
+import com.notably.model.block.Block;
+import com.notably.model.block.BlockTreeItem;
 import com.notably.model.suggestion.SuggestionItem;
 import com.notably.model.suggestion.SuggestionItemImpl;
 
@@ -29,40 +30,39 @@ public class OpenSuggestionCommand implements SuggestionCommand {
     @Override
     public void execute(Model model) {
         model.setResponseText(RESPONSE_MESSAGE);
-        System.out.println(RESPONSE_MESSAGE);
 
         List<AbsolutePath> possiblePaths = getPossiblePaths(path);
         List<SuggestionItem> suggestions = getSuggestions(possiblePaths, model);
 
-        for (SuggestionItem suggestionItem : suggestions) {
-            System.out.println(suggestionItem.getDisplayText());
-        }
+        //for (SuggestionItem suggestionItem : suggestions) {
+            //System.out.println(suggestionItem.getDisplayText());
+        //}
 
         model.setSuggestions(suggestions);
     }
 
     private List<AbsolutePath> getPossiblePaths(AbsolutePath path) {
-        Node<String> root = createTree();
+        BlockTreeItem root = getRootBlock();
 
         List<String> components = path.getComponents();
-        List<Node<String>> children = new ArrayList<>();
-        Queue<Node<String>> queue = new LinkedList<>();
+        List<BlockTreeItem> children = new ArrayList<>();
+        Queue<BlockTreeItem> queue = new LinkedList<>();
 
-        for (Node<String> rootChild : root.getChildren()) { //don't need add root to queue, but add root's children
+        for (BlockTreeItem rootChild : root.getBlockChildren()) { //don't need add root to queue, but add root's children
             queue.add(rootChild);
         }
 
         int index = 0;
         while (!queue.isEmpty()) {
-            Node<String> curr = queue.poll();
+            BlockTreeItem curr = queue.poll();
             String dir = components.get(index);
             if (curr.getTitle().equals(dir)) {
-                children = curr.getChildren();
+                children = curr.getBlockChildren();
                 if (index == components.size() - 1) { // just get all the children
                     break;
                 } else {
                     queue.clear();
-                    for (Node<String> child : children) {
+                    for (BlockTreeItem child : children) {
                         queue.add(child);
                     }
                     index++;
@@ -72,26 +72,26 @@ public class OpenSuggestionCommand implements SuggestionCommand {
 
         possiblePaths.add(path);
         if (!children.isEmpty()) {
-            for (Node<String> child : children) {
+            for (BlockTreeItem child : children) {
                 List<String> newComponents = new ArrayList<>();
                 newComponents.addAll(components);
-                newComponents.add(child.getTitle());
+                newComponents.add(child.getTitle().getText());
                 getChildRecursive(child, newComponents);
             }
         }
         return possiblePaths;
     }
 
-    private void getChildRecursive(Node<String> node, List<String> components) {
-        List<Node<String>> children = node.getChildren();
+    private void getChildRecursive(BlockTreeItem node, List<String> components) {
+        List<BlockTreeItem> children = node.getBlockChildren();
         if (children.size() == 0) { // if reach the end, add to possible paths
             AbsolutePath newPath = AbsolutePath.fromComponents(components);
             possiblePaths.add(newPath);
         } else {
-            for (Node<String> child : children) {
+            for (BlockTreeItem child : children) {
                 List<String> newComponents = new ArrayList<>();
                 newComponents.addAll(components);
-                newComponents.add(child.getTitle());
+                newComponents.add(child.getTitle().getText());
                 getChildRecursive(child, newComponents);
             }
         }
@@ -107,40 +107,5 @@ public class OpenSuggestionCommand implements SuggestionCommand {
                     return new SuggestionItemImpl(displayText, action);
                 })
                 .collect(Collectors.toList());
-    }
-
-    /**
-     * Creates a tree with dummy values. TODO: to be updated with the real DS.
-     * @return The root of the tree.
-     */
-    public static Node<String> createTree() {
-        Node<String> root = new Node<>("nus");
-
-        Node<String> nus = root.addChild(new Node<String>("nus"));
-        Node<String> y2s1 = nus.addChild(new Node<String>("y2s1"));
-        Node<String> y2s2 = nus.addChild(new Node<String>("y2s2"));
-        Node<String> cs2103t = y2s2.addChild(new Node<String>("cs2103t"));
-        Node<String> cs2101 = y2s2.addChild(new Node<String>("cs2101"));
-        Node<String> cs3243 = y2s2.addChild(new Node<String>("cs3243"));
-        Node<String> tp = cs2103t.addChild(new Node<String>("tp"));
-        Node<String> ip = cs2103t.addChild(new Node<String>("ip"));
-        Node<String> notably = tp.addChild(new Node<String>("notably"));
-
-        Node<String> personal = root.addChild(new Node<String>("personal"));
-        Node<String> p1 = personal.addChild(new Node<String>("p1"));
-        Node<String> p2 = personal.addChild(new Node<String>("p2"));
-
-        return root;
-    }
-
-    /**
-     * Prints the skeleton of the tree.
-     * @param node The root node.
-     * @param appender A separator to format the display of the tree.
-     * @param <T> A generic.
-     */
-    public static <T> void printTree(Node<T> node, String appender) {
-        System.out.println(appender + node.getTitle());
-        node.getChildren().forEach(each -> printTree(each, appender + appender));
     }
 }
